Create Question 3
For this question, we will keep track of when our friend’s birthdays are, and be able to find that information based on their name.
Create a dictionary of names and birthdays. When you run your program it should ask the user to enter a name, and return the birthday of that person back to them. (Note: You can store the bday in any format/datatype).